requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
apscheduler==3.10.4
pyyaml==6.0.2
httpx==0.27.2
anthropic==0.34.2
openai==1.45.0
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.8.2
python-dateutil==2.9.0
